+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/MetricUnionCompositeTable.java
@@ -12,44 +12,36 @@ import java.util.Set;
 import javax.validation.constraints.NotNull;
 
 /**
- * An implementation of <tt>BasePhysicalTable</tt> backed by metric union availability.
+ * An implementation of {@link PhysicalTable} backed by  {@link MetricUnionAvailability}.
  * <p>
- * The composite table puts metric columns of different tables together so that we can
- * query different metric columns from different tables at the same time.
+ * Under this composite, unioning is only legal if all metrics are uniquely sourced within the source tables.
  * <p>
- * For example, two tables of the following
+ * For example, given two source tables with metrics such that
  * <pre>
  * {@code
  * table1:
  * +---------+---------+---------+
  * | metric1 | metric2 | metric3 |
  * +---------+---------+---------+
- * |         |         |         |
- * |         |         |         |
- * +---------+---------+---------+
  *
  * table2
  * +---------+---------+---------+
- * | metric4 | metric5 | metric5 |
- * +---------+---------+---------+
- * |         |         |         |
- * |         |         |         |
+ * | metric4 | metric5 | metric6 |
  * +---------+---------+---------+
  * }
  * </pre>
- * are put together into a table
+ * the metric schema of this composite table will look like
  * <pre>
  * {@code
  * +---------+---------+---------+---------+---------+---------+
- * | metric1 | metric2 | metric3 | metric4 | metric5 | metric5 |
- * +---------+---------+---------+---------+---------+---------+
- * |         |         |         |         |         |         |
- * |         |         |         |         |         |         |
+ * | metric1 | metric2 | metric3 | metric4 | metric5 | metric6 |
  * +---------+---------+---------+---------+---------+---------+
  * }
  * </pre>
- * and this joined table is backed by the <tt>MetricUnionAvailability</tt>
+ * The available times are based on overlapping availability of participating sources provided by
+ * {@link MetricUnionAvailability}.
  *
+ * @see MetricUnionAvailability
  */
 public class MetricUnionCompositeTable extends BaseCompositePhysicalTable {