+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/time/ZonedTimeGrainSpec.groovy
@@ -21,6 +21,8 @@ class ZonedTimeGrainSpec extends Specification {
     static final String DENVER_NAME = "America/Denver"
     static final DateTimeZone DENVER_TZ = DateTimeZone.forID(DENVER_NAME)
 
+    public static final DAY_UTC = new ZonedTimeGrain(DAY, UTC)
+
     @Unroll
     def "test time zone granularity withZone turns #original into #expected"() {
         expect:
@@ -28,7 +30,7 @@ class ZonedTimeGrainSpec extends Specification {
 
         where:
         original                      | expected
-        new ZonedTimeGrain(DAY, UTC)  | new ZonedTimeGrain(DAY, DENVER_TZ)
+        DAY_UTC                       | new ZonedTimeGrain(DAY, DENVER_TZ)
         new ZonedTimeGrain(WEEK, UTC) | new ZonedTimeGrain(WEEK, DENVER_TZ)
     }